Dirty hack to fix deadstoping on mesh colliders. Need to look into a better way.
Fix is done by first seeing if we failed to move as far as we wanted, if we didn't and we were moving horizontally, then we try shifting left or right from our original position, and if that is open we try moving from there. In cases where this results in a noticable increase in movement we use that movement instead.